Move RTL helpers to `src/test/rtl` instead of being exported by `src/test/`
- Since RTL is a devDependency not a peerDependency, we can't assume consumers will have the library installed, and need to make reaching for it optional
