+++ src/test/rtl/component_helpers.ts
