+++ src/test/rtl/index.ts
@@ -6,12 +6,6 @@
  * Side Public License, v 1.
  */
 
-export { requiredProps } from './required_props';
-export { takeMountedSnapshot } from './take_mounted_snapshot';
-export { findTestSubject } from './find_test_subject';
-export {
-  startThrowingReactWarnings,
-  stopThrowingReactWarnings,
-} from './react_warnings';
-export { sleep } from './sleep';
-export * from './emotion-prefix';
+export * from './component_helpers';
+export * from './data_test_subj_queries';
+export { render, screen } from './custom_render';
